+++ app/lib/image-sanitizer.server.ts
@@ -1,13 +1,8 @@
 import sharp, { type Sharp } from 'sharp'
-
-export const SANITIZABLE_IMAGE_TYPES = [
-	'image/jpeg',
-	'image/png',
-	'image/webp',
-	'image/gif',
-] as const
-
-export type SanitizableImageType = (typeof SANITIZABLE_IMAGE_TYPES)[number]
+import {
+	isSanitizableImageType,
+	type SanitizableImageType,
+} from '~/lib/image-types'
 
 export type SanitizedImage = {
 	buffer: Buffer
@@ -38,12 +33,6 @@ const IMAGE_OUTPUTS: Record<
 		extension: 'gif',
 		encode: (image) => image.gif(),
 	},
-}
-
-export function isSanitizableImageType(
-	contentType: string,
-): contentType is SanitizableImageType {
-	return SANITIZABLE_IMAGE_TYPES.includes(contentType as SanitizableImageType)
 }
 
 export async function sanitizeImageFile(file: File): Promise<SanitizedImage> {
